minor: switch np.int to np.int32

+++ pymarchenko/marchenko.py
@@ -291,7 +291,7 @@ class Marchenko():
         """
         # Create window
         trav_off = trav - self.toff
-        trav_off = np.round(trav_off / self.dt).astype(np.int)
+        trav_off = np.round(trav_off / self.dt).astype(np.int32)
 
         w = np.zeros((self.nr, self.nt), dtype=self.dtype)
         for ir in range(self.nr):
@@ -459,7 +459,7 @@ class Marchenko():
 
         # Create window
         trav_off = trav - self.toff
-        trav_off = np.round(trav_off / self.dt).astype(np.int)
+        trav_off = np.round(trav_off / self.dt).astype(np.int32)
 
         w = np.zeros((self.nr, nvs, self.nt), dtype=self.dtype)
         for ir in range(self.nr):
